move custom accent modal to editor settings modal
