--- src/components/nb-custom-accent-modal/file-input.css
@@ -1,16 +1,0 @@
-@import "../../css/colors.css";
-
-.container {
-    display: block;
-    position: relative;
-    padding: 1rem;
-    text-align: center;
-    border: 4px dashed $ui-tertiary;
-    border-radius: 5px;
-    cursor: pointer;
-    background: none;
-    width: 100%;
-}
-.container:focus-within {
-    border-color: $looks-secondary;
-}
