+++ src/components/nb-editor-settings-modal/editor-settings-modal.css
@@ -8,7 +8,7 @@
 
 .body {
     background: $ui-white;
-    max-height: calc(100vh - 100px - 3.125rem);
+    height: calc(100vh - 100px - 3.125rem);
     overflow: auto;
     display: flex;
 }
@@ -263,7 +263,9 @@ label {
 
 .topic-list {
     width: 30%;
+    flex-shrink: 0;
     border-right: 1px solid $ui-black-transparent;
+    height: 100%;
 }
     
 .topic-item {
@@ -285,4 +287,6 @@ label {
 
 .content {
     padding: 1.5rem 2.25rem;
+    height: 100%;
+    overflow: auto;
 }
