+++ src/reducers/modals.js
@@ -19,7 +19,6 @@ const MODAL_RESTORE_POINTS = 'restorePointModal';
 const MODAL_FONTS = 'fontsModal';
 const MODAL_UNKNOWN_PLATFORM = 'unknownPlatformModal';
 const MODAL_INVALID_PROJECT = 'invalidProjectModal';
-const MODAL_CUSTOM_ACCENT = 'customAccentModal';
 const MODAL_EDITOR_SETTINGS = 'editorSettingsModal';
 
 const initialState = {
@@ -41,7 +40,6 @@ const initialState = {
     [MODAL_FONTS]: false,
     [MODAL_UNKNOWN_PLATFORM]: false,
     [MODAL_INVALID_PROJECT]: false,
-    [MODAL_CUSTOM_ACCENT]: false,
     [MODAL_EDITOR_SETTINGS]: false
 };
 
@@ -126,9 +124,6 @@ const openUnknownPlatformModal = function () {
 const openInvalidProjectModal = function () {
     return openModal(MODAL_INVALID_PROJECT);
 };
-const openCustomAccentModal = function () {
-    return openModal(MODAL_CUSTOM_ACCENT);
-};
 const openEditorSettingsModal = function () {
     return openModal(MODAL_EDITOR_SETTINGS);
 };
@@ -186,9 +181,6 @@ const closeUnknownPlatformModal = function () {
 const closeInvalidProjectModal = function () {
     return closeModal(MODAL_INVALID_PROJECT);
 };
-const closeCustomAccentModal = function () {
-    return closeModal(MODAL_CUSTOM_ACCENT);
-};
 const closeEditorSettingsModal = function () {
     return closeModal(MODAL_EDITOR_SETTINGS);
 };
@@ -213,7 +205,6 @@ export {
     openFontsModal,
     openUnknownPlatformModal,
     openInvalidProjectModal,
-    openCustomAccentModal,
     openEditorSettingsModal,
     closeBackdropLibrary,
     closeCostumeLibrary,
@@ -233,6 +224,5 @@ export {
     closeFontsModal,
     closeUnknownPlatformModal,
     closeInvalidProjectModal,
-    closeCustomAccentModal,
     closeEditorSettingsModal
 };
